requirements.txt
django

djangorestframework

django-cors-headers
djangorestframework-simplejwt

django-htmx
# NOTE: removed invalid/mis-typed package line that caused pip install to fail
# (original file contained a concatenated name 'python-decouplerequests==2.32.5')

# Geocoding helper
geopy>=2.0

# Optional: Google Maps geocoding client
googlemaps>=4.5


asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
geographiclib==2.1
geopy==2.4.1
idna==3.11
requests==2.32.5
sqlparse==0.5.4
urllib3==2.6.2
googlemaps==4.5.3

